1. Added new methods storeData() and clearData() in IDataStore. 2. Moved Attribute and FieldType class to text-api project. 3. Removed SampleScanBasedOperator and added LuceneDataReader to test LuceneDataStore class
