+++ textdb/textdb-api/src/main/java/edu/uci/ics/textdb/api/common/Attribute.java
@@ -1,4 +1,4 @@
-package edu.uci.ics.textdb.common.field;
+package edu.uci.ics.textdb.api.common;
 
 public class Attribute {
     private String fieldName;
