+++ textdb/textdb-common/src/main/java/edu/uci/ics/textdb/common/utils/Utils.java
@@ -9,10 +9,10 @@ import org.apache.lucene.document.Field.Store;
 import org.apache.lucene.document.IntField;
 import org.apache.lucene.index.IndexableField;
 
+import edu.uci.ics.textdb.api.common.FieldType;
 import edu.uci.ics.textdb.api.common.IField;
 import edu.uci.ics.textdb.common.field.DateField;
 import edu.uci.ics.textdb.common.field.DoubleField;
-import edu.uci.ics.textdb.common.field.FieldType;
 import edu.uci.ics.textdb.common.field.IntegerField;
 import edu.uci.ics.textdb.common.field.StringField;
