+++ textdb/textdb-storage/src/test/java/edu/uci/ics/textdb/storage/LuceneDataStoreTest.java
@@ -7,11 +7,10 @@ import junit.framework.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
+import edu.uci.ics.textdb.api.common.Attribute;
 import edu.uci.ics.textdb.api.common.ITuple;
-import edu.uci.ics.textdb.api.dataflow.IOperator;
 import edu.uci.ics.textdb.common.constants.LuceneConstants;
 import edu.uci.ics.textdb.common.constants.TestConstants;
-import edu.uci.ics.textdb.common.field.Attribute;
 
 public class LuceneDataStoreTest {
     private LuceneDataStore dataStore;
@@ -28,16 +27,11 @@ public class LuceneDataStoreTest {
         List<ITuple> tuples = TestConstants.getSamplePeopleTuples();
         dataStore.storeData(schema, tuples);
         
-        //Using SampleScanBased Operator since we cannot use ScanBasedSourceOperator present in textdb-dataflow project.
-        //If textdb-storage references textdb-dataflow it creates a cyclic dependency.
-        IOperator operator = new SampleScanBasedOperator(LuceneConstants.INDEX_DIR, TestConstants.SAMPLE_SCHEMA_PEOPLE);
-        operator.open();
-        ITuple tuple = null;
-        int numTuples = 0;
-        while((tuple = operator.getNextTuple()) != null){
-            numTuples++;
-        }
-        Assert.assertEquals(tuples.size(), numTuples);
-        operator.close();
+        LuceneDataReader luceneDataReader = new LuceneDataReader(
+                LuceneConstants.INDEX_DIR, TestConstants.SAMPLE_SCHEMA_PEOPLE);
+        List<ITuple> tuplesFetched = luceneDataReader.getTuples();
+
+        Assert.assertEquals(tuples.size(), tuplesFetched.size());
+        
     }
 }
